[FLINK-16357][checkpointing] Offer different methods for "global restore" and "local restore" in CheckpointCoordinator
Global restores are meant for initial savepoint restores, and for global failover in the scheduler.
Global failover in the scheduler happens for example during master failover and as a safety net when
encountering unexpected/inconsistent situations that might impact correctness.

Local failovers are all common task failures and recoveries.

This commit offers different methods to be called in these two situations, but does not make use of the
local restore method yet. All calls still go to the global restore, which was the previous behavior in
all cases.

The difference in the CheckpointCoordinator between local and global restore is currently that OperatorCoordinators
are only restored during global restores.

As a side effect, this change also eliminates the "failWhenNoCheckpoint" flag outside of the CheckpointCoordinator.
The flag is exclusively used by the "restoreSavepoint()" method which is a separate call to the Coordinator anyways.