[FLINK-17541][table] Support inline structured types
This enables inline structured types in the Blink planner. Inline
structured types are extracted (e.g. in UDFs) and don't need to be
registered in a catalog. This finalizes FLIP-65 for scalar and
table functions because existing functions can be migrated with a
replacement to the new type system.

Structured type support should still be declared as experimental
until we have more tests and can also deal with structured types
in sources and sinks.

This closes #12228.