[FLINK-15794][Kubernetes] Generate the Kubernetes default image version
The default image used by Kubernetes is 'flink:latest' which causes version compatibility problems
if the latest it not exactly the same as what you are using.
The commit derives the default value from the actual Flink and Scala version,
which the running Flink was built with (i.e. no longer latest).
The latest tag is used only for snapshot versions until we have snapshot builds for docker images.

This closes #11245.