[FLINK-18739] Implement MergingSharedSlotProfileRetriever
Input location preferences will be considered for each SharedSlot when allocating a physical slot for it.
Input location preferences of a SharedSlot are the merged input location preferences of all the tasks to run in this SharedSlot.

Inter-ExecutionSlotSharingGroup input location preferences can be respected in this way for ExecutionSlotSharingGroups belonging to different bulks.
If ExecutionSlotSharingGroups belong to the same bulk, the input location preferences are ignored because of possible cyclic dependencies.
Later, we can optimise this case when the declarative resource management for reactive mode is ready.

Intra-ExecutionSlotSharingGroup input location preferences will also be respected when creating ExecutionSlotSharingGroup(s) in LocalInputPreferredSlotSharingStrategy.

This closes #13028.